alterado int por float
alterado tipo na declaração das variaveis

+++ AlgLog in C-C++/ex001.cpp
@@ -5,7 +5,7 @@ int main() {
   // 1) Faça um programa que receba três notas, calcule e mostre a média
   // aritmética entre elas.//
 
-  int n1, n2, n3, media;
+  float n1, n2, n3, media;
   cout << "Insira a nota 1: ";
   cin >> n1;
   cout << "Insira a nota 2: ";
